Move components into folder + Update paths
