+++ packages/app/src/components/App.js
@@ -4,13 +4,13 @@ import PropTypes from 'prop-types';
 import Snackbar from '@material-ui/core/Snackbar';
 import i18n from '@dhis2/d2-i18n';
 
-import SnackbarMessage from './widgets/SnackbarMessage';
+import SnackbarMessage from '../widgets/SnackbarMessage';
 import MenuBar from './MenuBar/MenuBar';
 import VisualizationTypeSelector from './VisualizationTypeSelector/VisualizationTypeSelector';
 import Dimensions from './Dimensions/Dimensions';
 import Visualization from './Visualization/Visualization';
-import * as fromReducers from './reducers';
-import * as fromActions from './actions';
+import * as fromReducers from '../reducers';
+import * as fromActions from '../actions';
 
 import './App.css';
