+++ packages/app/src/components/Dimensions/Dimensions.js
@@ -2,7 +2,7 @@ import React, { Component } from 'react';
 import { TextField } from '@material-ui/core';
 import { Search } from '@material-ui/icons';
 import i18n from '@dhis2/d2-i18n';
-import { colors } from '../colors';
+import { colors } from '../../colors';
 import DimensionList from './DimensionList';
 import DialogManager from './DialogManager';
