+++ packages/app/src/components/Dimensions/__tests__/DialogManager.spec.js
