+++ packages/app/src/components/VisualizationOptions/StyleTab.js
