+++ packages/app/src/components/VisualizationOptions/__tests__/AxesAndLegendsTab.spec.js
