+++ packages/app/src/components/VisualizationOptions/__tests__/DataTabCheckBoxes.spec.js
