+++ packages/app/src/components/VisualizationOptions/__tests__/StyleTab.spec.js
