+++ packages/app/src/components/VisualizationOptions/__tests__/TabsBar.spec.js
